Move node's new analyzer to separate go file
